feat(CHROME-1): dev/prod webpack config added
1. development webpack config added
2. production webpack config added
