+++ tools/production.js
@@ -1,0 +1,14 @@
+/**
+ * production configuration
+ *
+ * @author kern86
+ * @since 2019.04.04 - draft
+ */
+'use strict';
+
+const merge = require('webpack-merge');
+const base = require('./base');
+
+module.exports = merge(base, {
+  mode: 'production',
+});
